refactor: split agent task dispatch command
