+++ src/commands/mod.rs
@@ -166,6 +166,7 @@ pub fn finalize_set_spec(
 }
 
 pub mod agent_task;
+pub(crate) mod agent_task_dispatch;
 pub mod api;
 pub mod audit;
 pub mod auth;
